separate solver class from bmc_tool.py
